0601 jo 답변 에디터 구현중.

+++ yedam_finalProject/src/main/java/com/yedam/finalPrj/product/vo/park/hong/ProductReservationVO.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class ProductReservationVO {
 	
 //  상품 예약
-	private long prodResNo;
+	private int prodResNo;
 	private int storeNo;
 	private int memNo;
 	private Date pickupTime;
